Move the overlap offset geometry out of WindowManager
It has no dependency on WindowManager's state, so it does not need to
live there.
